adding octet-stream mp4 support
Sometimes when downloading a live stream with yt-dlp it's interrupted or for whatever reason doesn't format correctly and the video file has a mp4 extension but is still in the MPEG-TS format so video/* won't open it. By adding mp4 to the octet-stream ts case now the videos will open and play as they are and don't require you to rename them with a .ts extension or to convert them with ffmpeg.